+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 ### Added
-- Boostrap cards snippet
+- Bootstrap cards snippet
+- Bootstrap progress-bar snippet
 
 ## [v0.0.1]
 ### Added